Added majority voting for each fold combining the top 5 models

+++ tests/test_moses_cross_val.py
@@ -7,7 +7,7 @@ import uuid
 from crossval.moses_cross_val import CrossValidation
 import pandas as pd
 import glob
-
+from sklearn.model_selection import train_test_split
 
 class TestCrossValidation(unittest.TestCase):
 
@@ -15,6 +15,23 @@ class TestCrossValidation(unittest.TestCase):
         dataset = os.path.join(TEST_DATA_DIR, "bin_truncated.csv")
         session_id = str(uuid.uuid4())
         self.session = Session(session_id, moses_options, crossval_options, dataset)
+
+        self.train_file, self.test_file = os.path.join(TEST_DATA_DIR, "train_file"), os.path.join(TEST_DATA_DIR, "test_file")
+        self.test_matrix = [[1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1],
+                            [1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1],
+                            [0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 0, 0, 0, 1],
+                            [1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1],
+                            [1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 0, 1],
+                            [0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1],
+                            [0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 0, 0, 1],
+                            [0, 1, 0, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 1, 1, 1, 0, 0, 1],
+                            [0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1],
+                            [1, 1, 1, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1],
+                            [1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1],
+                            [1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 0, 0, 1, 1, 1, 0, 1],
+                            [0, 1, 0, 0, 1, 1, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0, 1, 1, 0, 0, 0, 1],
+                            [1, 1, 1, 1, 1, 0, 1, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 1, 0, 1],
+                            [1, 1, 1, 1, 1, 1, 1, 0, 0, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 1, 1, 0, 1]]
 
     def test_run_folds(self):
         moses_cross_val = CrossValidation(self.session, None, TEST_DATA_DIR)
@@ -29,10 +46,25 @@ class TestCrossValidation(unittest.TestCase):
 
         self.assertEqual(len(fold_df.columns), 12)
 
+    def test_majority_vote(self):
+        df = pd.read_csv(self.session.dataset)
+
+        train_df, test_df = train_test_split(df, test_size=0.3)
+
+        test_df.to_csv(self.test_file, index=False)
+
+        moses_cross_val = CrossValidation(self.session, None, TEST_DATA_DIR)
+        moses_cross_val.test_file = self.test_file
+
+        scores = moses_cross_val.majority_vote(self.test_matrix)
+
+        self.assertEqual(scores.shape, (1, 5))
+
     def tearDown(self):
-        os.remove("test_file")
-        os.remove("train_file")
-        os.remove("feature_count.csv")
+        if os.path.exists(self.test_file): os.remove(self.test_file)
+        if os.path.exists(self.train_file): os.remove(self.train_file)
+        if os.path.exists("feature_count.csv"): os.remove("feature_count.csv")
+
         moses_log = os.path.join(os.path.dirname(__file__), "moses.log")
         opencog_log = os.path.join(os.path.dirname(__file__), "opencog.log")
         if os.path.exists(moses_log):
@@ -41,6 +73,9 @@ class TestCrossValidation(unittest.TestCase):
             os.remove(opencog_log)
 
         for file in glob.glob("fold_[0-9].csv"):
+            os.remove(file)
+
+        for file in glob.glob("ensemble_[0-9].csv"):
             os.remove(file)
